Issue 5245: Improve extent management
git-svn-id: http://argouml.tigris.org/svn/argouml/trunk@15410 a161b567-7d1e-0410-9ef9-912c70fedb3f

+++ src/argouml-app/src/org/argouml/kernel/ProjectImpl.java
@@ -46,6 +46,7 @@ import org.argouml.application.api.Argo;
 import org.argouml.application.helpers.ApplicationVersion;
 import org.argouml.configuration.Configuration;
 import org.argouml.i18n.Translator;
+import org.argouml.model.InvalidElementException;
 import org.argouml.model.Model;
 import org.argouml.persistence.PersistenceManager;
 import org.argouml.profile.Profile;
@@ -1067,8 +1068,14 @@ public class ProjectImpl implements java.io.Serializable, Project {
         }
 
         members.clear();
-        Model.getUmlFactory().deleteExtent(roots.iterator().next());
-        roots.clear();
+        if (!roots.isEmpty()) {
+            try {
+                Model.getUmlFactory().deleteExtent(roots.iterator().next());
+            } catch (InvalidElementException e) {
+                LOG.warn("Extent deleted a second time");
+            }
+            roots.clear();
+        }
         models.clear();
         diagrams.clear();
         searchpath.clear();
